+++ main.go
@@ -10,17 +10,18 @@ import (
 
 const version string = "1.2.0"
 
+// command line flags
 var showVersion bool
 var quietProcess bool
 
-var stderr = os.Stderr
-
+// command structure
 type cliCmd struct {
 	CmdName string
 	CmdDesc string
 	CmdFunc func(Chronometer, string)
 }
 
+// list of permitted commands with description and func pointer
 var cmdList = []cliCmd{
 	cliCmd{"start", "start timer", Chronometer.start},
 	cliCmd{"read", "read timer (elapsed time)", Chronometer.read},
@@ -30,8 +31,8 @@ var cmdList = []cliCmd{
 	cliCmd{"exec", "execute process and print elapsed time", Chronometer.exec},
 }
 
-// define flags in init() otherwise you'll get "panic: flag redefined"
-// when calling main twice while testing.
+// Note: defining flags in init() to avoid "panic: flag redefined"
+// when calling main() twice in main_test.
 func init() {
 	flag.BoolVar(&showVersion, "version", false, "print version and exit")
 	flag.BoolVar(&quietProcess, "quiet", false, "hide process output")
@@ -48,42 +49,44 @@ func main() {
 	}
 }
 
-// If flag not specified in first position, it will be ignored.
-// Example: ./timer start azerty -C ls
-// --> flag '-C' will be ignored
+// Note: if flag not specified in first position, it will be ignored.
+// Example: ./timer start azerty -C ls --> flag '-C' will be ignored
 func processArgs(cmd string, name string) {
-	setUsage()
+	setCustomUsage()
 	checkParameters(cmd, name)
 	executeTimerFunc(cmd, name)
 }
 
-func setUsage() {
+// Redefines flag.Usage() function.
+func setCustomUsage() {
 	tpl, _ := template.New("usage").Parse(`{{range .}}  {{.CmdName}}: {{.CmdDesc}}
 {{end}}`)
 	flag.Usage = func() {
-		fmt.Fprintf(stderr, "Usage: %s [option] command name\n", os.Args[0])
-		fmt.Fprintf(stderr, "\n COMMANDS:\n")
-		tpl.Execute(stderr, cmdList)
-		fmt.Fprintf(stderr, "\n OPTIONS:\n")
+		fmt.Fprintf(os.Stderr, "Usage: %s [option] command name\n", os.Args[0])
+		fmt.Fprintf(os.Stderr, "\n COMMANDS:\n")
+		tpl.Execute(os.Stderr, cmdList)
+		fmt.Fprintf(os.Stderr, "\n OPTIONS:\n")
 		flag.PrintDefaults()
 		os.Exit(1)
 	}
 }
 
+// Verifies command line parameters.
 func checkParameters(cmd string, name string) {
 	if flag.NArg() == 0 || flag.NArg() > 2 {
 		flag.Usage()
 	}
 	if (cmd == "start" || cmd == "read" || cmd == "stop") && name == "" {
-		fmt.Fprintf(stderr, "Please specify the name of a timer.\n")
+		fmt.Fprintf(os.Stderr, "Please specify the name of a timer.\n")
 		os.Exit(1)
 	}
 	if cmd == "exec" && name == "" {
-		fmt.Fprintf(stderr, "Please specify a process to execute.\n")
+		fmt.Fprintf(os.Stderr, "Please specify a process to execute.\n")
 		os.Exit(1)
 	}
 }
 
+// Executes the requested action defined in command list.
 func executeTimerFunc(cmd string, name string) {
 	i := indexOf(cmd)
 	if i != -1 {
@@ -93,8 +96,8 @@ func executeTimerFunc(cmd string, name string) {
 	}
 }
 
-// check specified command against supported commands
-// and return index if found, -1 otherwise
+// Checks specified command against supported commands
+// and return index if found, -1 otherwise.
 func indexOf(cmdName string) int {
 	for i, item := range cmdList {
 		if item.CmdName == cmdName {