+++ CHANGELOG.md
@@ -3,8 +3,8 @@
 ##### timer 1.2.0 - (14 July 2015)
 
     * Replaced flags by commands
+    * Stopping timer before accessing registry
     * More consistent log output
- 
  
 ##### timer 1.1.1 - (11 June 2015)